# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NantonFilm API - proxy to external streaming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProxyResponse {
  ok: boolean;
  data?: unknown;
}

export type SearchContentParams = {
  q: string;
  page?: number;
};

export type GetAnimeDonghuaParams = {
  page?: number;
};

export type GetFilmParams = {
  page?: number;
};

export type GetSeriesParams = {
  page?: number;
};

export type GetTvshowParams = {
  page?: number;
};

export type GetOthersParams = {
  page?: number;
};

export type GetGenreDetailParams = {
  page?: number;
};

export type GetScheduleParams = {
  day?: GetScheduleDay;
};

export type GetScheduleDay =
  (typeof GetScheduleDay)[keyof typeof GetScheduleDay];

export const GetScheduleDay = {
  senin: "senin",
  selasa: "selasa",
  rabu: "rabu",
  kamis: "kamis",
  jumat: "jumat",
  sabtu: "sabtu",
  minggu: "minggu",
} as const;

export type GetAllAnimeParams = {
  page?: number;
};

export type GetCatalogParams = {
  title?: string;
  page?: number;
  order?: string;
  type?: string;
  status?: string;
};

export type GetServerParams = {
  post: string;
  nume: string;
  type: string;
};

export type GetProxyImageParams = {
  url: string;
};