# lib/api-zod/src/generated/types/proxyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NantonFilm API - proxy to external streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface ProxyResponse {
  ok: boolean;
  data?: unknown;
}